requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-ckeditor==6.1.0
django-froala-editor==3.2.2
django-js-asset==1.2.2
django-summernote==0.8.11.6
django-tinymce==3.3.0
django-tinymce4-lite==1.8.0
gunicorn==20.1.0
idna==2.10
jsmin==2.2.2
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
